book_depository_be/requirement.txt
django==3.2.5
djangorestframework==3.12.4
djongo==1.3.6
pillow==8.3.1
djangorestframework-jwt==1.11.0